docs: Add Python notes README
This commit adds a README file with some notes on how to use Python
for various tasks, such as testing code with doctest, choosing a license,
creating a gitignore file, and formatting strings with the
.format() method. The README also includes some examples of code and
output, using the Python console and doctest.
Some parts of the output are skipped or elided for brevity,
using doctest directives.